Activate gemspec after successfully requiring a gem
Ensure that `Gem.loaded_specs[gem_name]` contains the Gem::Specification
instance we locate and set `@activated` and `@loaded` on that spec to
`true`.

Rubygems does this automatically when it activates a gem so anything
that relies on any of the above would fail without this change.

This was specifically a bug when using the `meta_request` gem as it
contains this code:

  module MetaRequest
    VERSION = Gem.loaded_specs["meta_request"].version.to_s
  end